few errors fixed and changed the colors

+++ database/migrations/2026_02_02_135621_add_genre_id_to_books_table.php
@@ -14,13 +14,16 @@ return new class extends Migration
         Schema::table('books', function (Blueprint $table) {
 
     // Add genre_id column after status
-            $table->unsignedBigInteger('genre_id')->nullable()->after('status');
+
+  if (!Schema::hasColumn('books', 'genre_id')) {
+        $table->unsignedBigInteger('genre_id')->after('status');
 
             // Set foreign key relationship to genres table
             $table->foreign('genre_id')
                   ->references('id')
                   ->on('genres')
                   ->onDelete('set null'); // if genre is deleted, books keep genre_id null
+  }
         });
     }
